Created Index.html
Within the amount I have done so far I have laid out the foundation for more elements to be added with images to work on the layout and plan of the webpage using flexbox and css declarations. Will be working back and forth from home and work to make this repository better. 16/6/22